Factor out UnsavedChanges as a standalone module
