+++ frontend/src/modules/editor/components/Editor.test.js
@@ -24,6 +24,12 @@ const SELECTED_ENTITY = {
     ],
 };
 
+const UNSAVED_CHANGES = {
+    unsavedChanges: false,
+    unsavedChangesCallback: null,
+    unsavedChangesIgnored: false,
+}
+
 
 function createEditorBase({
     pluralForm = -1,
@@ -46,6 +52,7 @@ function createEditorBase({
                 forceSuggestions,
             },
         } }
+        unsavedchanges={ UNSAVED_CHANGES }
         disableAction={ sinon.spy() }
     />);
 }
